pyproject.toml
########################## POETRY CONFIGURATION ###############################
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "time_series_classification"
version = "0.1.0"
description = "A short description of the project."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

#################### DEPENDENCIES AND DEV DEPENDENCIES ########################
[tool.poetry.dependencies]
python = ">=3.11, <3.12"
clearml = "1.13.1"
numpy = "1.26.1"
pandas = "2.1.2"
scikit-learn = "1.3.2"
xgboost = "2.0.3"
seaborn = "0.13.1"
stac-add-time-features = "^0.1.0"
mljar-supervised = "^1.1.3"

[tool.poetry.scripts]
train = "src.train"

#################### TOOLS AND SCRIPTS CONFIGURATION ###########################
[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "N",    # pep8-naming
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "PD",   # pandas vet
    "NPY",  # NumPy-specific rules
]

######################## REPOSITORIES AND SOURCES ############################## 
[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "stac"
url = "https://artifact.tecnalia.com/artifactory/api/pypi/ai-catalogue-pypi-dev-local/simple"
priority = "supplemental"
